only if we got a Process object back

+++ cwltest/__init__.py
@@ -206,8 +206,9 @@ def run_test(
         _logger.error(test.get("doc"))
         # Kill and re-communicate to get the logs and reap the child, as
         # instructed in the subprocess docs.
-        process.kill()
-        outstr, outerr = process.communicate()
+        if process:
+            process.kill()
+            outstr, outerr = process.communicate()
         return TestResult(2, outstr, outerr, timeout, args.classname, "Test timed out")
     finally:
         if process is not None and process.returncode is None:
